# packages/castv2-server/lib/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "HTTP API for Google Cast enabled devices. This API is only intended to be used on your local network not for hosting on the public internet.",
    "version": "1.1.0",
    "title": "cast-web-api",
    "license": {
      "name": "GNU General Public License v3.0",
      "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
    }
  },
...
  "schemes": ["http"],
  "paths": {
...
    "/assistant/broadcast": {
      "post": {
        "tags": ["assistant"],
        "summary": "Broadcast a message through the Google Assistant",
        "description": "Message will be read out a loud by all Google Assistant devices on the local network.",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Message object, that should be broadcast",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Message"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/Response"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Assistant error",
            "schema": {
              "$ref": "#/definitions/Response"
            }
          }
        }
      }
    },
    "/assistant/command": {
      "post": {
        "tags": ["assistant"],
        "summary": "Broadcast a message through the Google Assistant",
        "description": "Message will be read out a loud by all Google Assistant devices on the local network.",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Message object, that should be broadcast",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Message"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/Response"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Assistant error",
            "schema": {
              "$ref": "#/definitions/Response"
            }
          }
        }
      }
    },
...
    "/callback": {
      "get": {
        "tags": ["callback"],
        "summary": "List all registered callbacks",
        "description": "Sends a http package to a http url as soon as a device status or link changes.",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Callback"
              }
            }
          }
        }
      },
      "post": {
        "tags": ["callback"],
        "summary": "Create a new callback",
        "description": "Create a new callback, identified by its url. If a callback with the same url already exists, it will be overwritten.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Callback object, that should be created",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Callback"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "repsonse": {
                  "type": "string",
                  "default": "ok"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      },
      "put": {
        "tags": ["callback"],
        "summary": "Update an existing callback",
        "description": "Update an existing callback, identified by its url. Only the settings will be changed.",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Callback object, only with the properties that need to be updated",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Callback"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "repsonse": {
                  "type": "string",
                  "default": "ok"
                },
                "url": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "tags": ["callback"],
        "summary": "Delete an existing callback",
        "description": "Delete an existing callback, identified by its url.",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Callback object, with id that should be deleted",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Callback"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "repsonse": {
                  "type": "string",
                  "default": "ok"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Response": {
      "type": "object",
      "properties": {
        "response": {
          "type": "string",
          "description": "Indicates if the command completed successfully",
          "enum": ["ok", "error"]
        },
        "error": {
          "type": "string",
          "description": "Error description"
        }
      }
    },
...
    "Callback": {
      "type": "object",
      "properties": {
        "url": {
          "type": "string",
          "description": "HTTP url to which packages will be send"
        },
        "settings": {
          "type": "object",
          "description": "To be detailed",
          "properties": {}
        }
      }
    },
...
    "Message": {
      "type": "object",
      "properties": {
        "message": {
          "description": "Message for the Goggle Assistant",
          "type": "string"
        }
      }
    }
  },
...
}